Simplifiying a bit the code
